third_party/WebKit/LayoutTests/http/tests/misc/webtiming-resolution-expected.txt

Verifies the minimum resolution is 100 microseconds.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shouldBeNearZeroOrOne < 1e-5 || Math.abs(shouldBeNearZeroOrOne - 1) < 1e-5 is true
PASS shouldBeNearZeroOrOne < 1e-5 || Math.abs(shouldBeNearZeroOrOne - 1) < 1e-5 is true
PASS successfullyParsed is true

TEST COMPLETE
